Move interface augment to it's own file
